requirements.txt
# Core RAG dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Vector store options
chromadb==0.5.23
faiss-cpu>=1.9.0

# Data processing
pandas>=2.2.0
openpyxl>=3.1.2
numpy>=1.26.0

# LLM and embeddings (Ollama)
ollama>=0.1.0

# API and web interface
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment and utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Optional: Web interfaces (uncomment if needed)
# streamlit>=1.30.0
# gradio>=4.13.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
